+++ tests/DebugEventDispatcherTest.php
@@ -35,6 +35,7 @@ final class DebugEventDispatcherTest extends TestCase
             }
         };
         $eventDispatcher = new DebugEventDispatcher(new EventDispatcher($listenerProvider), $logger);
+        $this->assertFalse($eventDispatcher->dispatched($event::class));
         $this->assertSame($event, $eventDispatcher->dispatch($event));
         $this->assertSame(1, $var);
         $this->assertCount(1, $logger->records);
@@ -43,5 +44,7 @@ final class DebugEventDispatcherTest extends TestCase
             "type" => $event::class,
             "event" => $event,
         ], $logger->records[0]);
+        $this->assertTrue($eventDispatcher->dispatched($event::class));
+        $this->assertFalse($eventDispatcher->dispatched($event::class, 2));
     }
 }